+++ spec/controllers/dorsale/flyboy/task_comments_controller_spec.rb
@@ -11,9 +11,22 @@ describe Dorsale::Flyboy::TaskCommentsController, type: :controller do
   let(:task_comment) { create(:flyboy_task_comment, task: task) }
 
   describe "#create" do
+    let(:valid_params) {{task_id: task.id, task_comment: task_comment.attributes}}
+
     it "should create the task_comment" do
-      post :create, params: {task_id: task.id, task_comment: task_comment.attributes}
+      post :create, params: valid_params
       expect(assigns(:task_comment).persisted?).to be true
+    end
+
+    it "should redirect to referrer" do
+      request.env["HTTP_REFERER"] = "/abc"
+      post :create, params: valid_params
+      expect(response).to redirect_to "/abc"
+    end
+
+    it "should redirect to task" do
+      post :create, params: valid_params
+      expect(response).to redirect_to flyboy_task_path(task)
     end
   end